it does change the svg colour into a specified colour, but when clicked again it does not take it back into original colour
