+++ src/components/displaySvg/finalSvgDisplay.scss
@@ -9,10 +9,6 @@
     }
  }
 
- rect:hover {
-    background-color: yellow;
-    fill:green($color: #000000);
-  }
 
   .className-1{
       :hover{
@@ -23,9 +19,7 @@
       
   }
 
-  a:hover{
-     fill:yellow;
-  }
+
